requirements.txt
# Python 3.12.x
fastapi>=0.128.0
uvicorn[standard]>=0.40.0
python-dotenv>=1.2.1
pydantic>=2.12.5
pydantic-settings>=2.12.0
openai>=2.15.0
anthropic>=0.76.0
python-multipart>=0.0.22
python-docx>=1.2.0
reportlab>=4.2.0
